backend/requirements.txt
Django==5.2.2
psycopg[binary]==3.2.10
djangorestframework==3.16.0
python-dotenv==1.1.0
